Reorganize kafka tests
