+++ stdlib/messaging/kafka/src/test/resources/test-src/compiler-validation/custom_error_return_type_validation.bal
