+++ stdlib/messaging/kafka/src/test/resources/test-src/compiler-validation/invalid_resource_name.bal
