+++ stdlib/messaging/kafka/src/test/resources/test-src/compiler-validation/more_than_one_resource.bal
