+++ stdlib/messaging/kafka/src/test/resources/test-src/consumer/topic_subscribe_with_partition_rebalance.bal
