+++ stdlib/messaging/kafka/src/test/resources/test-src/services/advanced_service.bal
