+++ stdlib/messaging/kafka/src/test/resources/test-src/services/simple_service.bal
