+++ stdlib/messaging/kafka/src/test/resources/test-src/services/stop_service.bal
