+++ stdlib/messaging/kafka/src/test/resources/testng.xml
@@ -23,50 +23,50 @@
     <test name="ballerina-kafka-consumer-tests">
         <!-- Consumer-Related tests -->
         <classes>
-            <class name="org.ballerinalang.messaging.kafka.consumer.KafkaConsumerAssignTest" />
-            <class name="org.ballerinalang.messaging.kafka.consumer.KafkaConsumerTopicsTest" />
-            <class name="org.ballerinalang.messaging.kafka.consumer.KafkaConsumerTest" />
-            <class name="org.ballerinalang.messaging.kafka.consumer.KafkaConsumerManualCommitTest" />
-            <class name="org.ballerinalang.messaging.kafka.consumer.KafkaConsumerPauseResumeTest" />
-            <class name="org.ballerinalang.messaging.kafka.consumer.KafkaConsumerSubscribePartitionRebalanceTest" />
-            <class name="org.ballerinalang.messaging.kafka.consumer.KafkaConsumerSubscribeToPatternTest" />
-            <class name="org.ballerinalang.messaging.kafka.consumer.KafkaConsumerSeekTest" />
+            <class name="org.ballerinalang.messaging.kafka.consumer.TopicAssignTest" />
+            <class name="org.ballerinalang.messaging.kafka.consumer.TopicsTest" />
+            <class name="org.ballerinalang.messaging.kafka.consumer.ConsumerFunctionsTest" />
+            <class name="org.ballerinalang.messaging.kafka.consumer.ManualCommitTest" />
+            <class name="org.ballerinalang.messaging.kafka.consumer.TopicPauseResumeTest" />
+            <class name="org.ballerinalang.messaging.kafka.consumer.TopicSubscribePartitionRebalanceTest" />
+            <class name="org.ballerinalang.messaging.kafka.consumer.TopicSubscribeToPatternTest" />
+            <class name="org.ballerinalang.messaging.kafka.consumer.PartitionSeekTest" />
         </classes>
     </test>
     <test name="ballerina-kafka-producer-tests">
         <!-- Producer-Related tests -->
         <classes>
-            <class name="org.ballerinalang.messaging.kafka.producer.KafkaProducerTest" />
+            <class name="org.ballerinalang.messaging.kafka.producer.ProducerFunctionsTest" />
         </classes>
     </test>
     <test name="ballerina-kafka-service-tests">
         <!-- Service-Related tests -->
         <classes>
-            <class name="org.ballerinalang.messaging.kafka.services.KafkaServiceTest" />
+            <class name="org.ballerinalang.messaging.kafka.services.ServiceTest" />
         </classes>
     </test>
     <test name="ballerina-kafka-compiler-test">
         <!-- Compiler plugin validation tests -->
         <classes>
-            <class name="org.ballerinalang.messaging.kafka.compiler.KafkaServiceCompilerTest" />
+            <class name="org.ballerinalang.messaging.kafka.compiler.CompilerPluginTest" />
         </classes>
     </test>
     <test name="ballerina-kafka-transactions-test">
         <classes>
             <!-- Transactions-Related tests -->
-            <class name="org.ballerinalang.messaging.kafka.transactions.KafkaProducerTransactionsTest" />
+            <class name="org.ballerinalang.messaging.kafka.transactions.TransactionalProducerTest" />
         </classes>
     </test>
-    <test name="ballerina-kafka-ssl-tests">
+    <test name="ballerina-kafka-security-tests">
         <classes>
             <!-- SSL-Based Consumer/Producer-Related tests -->
-            <class name="org.ballerinalang.messaging.kafka.ssl.KafkaConsumerAndProducerWithSSLTest" />
+            <class name="org.ballerinalang.messaging.kafka.security.ConsumerAndProducerWithSslTest" />
         </classes>
     </test>
     <test name="ballerina-kafka-serializer-deserializer-tests">
         <classes>
             <!-- SSL-Based Consumer/Producer-Related tests -->
-            <class name="org.ballerinalang.messaging.kafka.serdes.KafkaSerializerDeserializerTest" />
+            <class name="org.ballerinalang.messaging.kafka.serdes.SerializerDeserializerTest" />
         </classes>
     </test>
 </suite>
